feat: convert tvoc to voc density
This commit adds support for converting TVOC to VOC Density, as tracked
by HomeKit. This conversion does require giving the target molecular
weight as input, though. Formeldahyde was picked as the default gas
weight since it's a common household gas, but we let the user configure
the gas weight just in case they have a better idea of what they want to
track.

For a TVOC of 300 ppb at 25 C and using the weight of formeldahyde, we
get 368 μg/m^3.